requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.17
langchain-openai==0.0.8
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0